requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.1.0
alembic==1.13.2
marshmallow==3.22.0
flask-smorest==0.44.0
flask-jwt-extended==4.6.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
PyYAML==6.0.2
celery==5.4.0
redis==5.0.8
gunicorn==22.0.0